fix(route/cna): Unify full text fetching method
