+++ lib/routes/cna/web/index.ts
@@ -4,6 +4,7 @@ import got from '@/utils/got';
 import { load } from 'cheerio';
 import { parseDate } from '@/utils/parse-date';
 import timezone from '@/utils/timezone';
+import { getFullText } from '../utils';
 
 export const route: Route = {
     path: '/web/:id?',
@@ -45,28 +46,7 @@ async function handler(ctx) {
             };
         });
 
-    const items = await Promise.all(
-        list.map((item) =>
-            cache.tryGet(item.link, async () => {
-                const detailResponse = await got({
-                    method: 'get',
-                    url: item.link,
-                });
-                const content = load(detailResponse.data);
-                const topImage = content('.fullPic').html();
-
-                item.description = (topImage === null ? '' : topImage) + content('.paragraph').eq(0).html();
-                item.category = [
-                    ...content("meta[property='article:tag']")
-                        .toArray()
-                        .map((e) => e.attribs.content),
-                    content('.active > a').text(),
-                ];
-
-                return item;
-            })
-        )
-    );
+    const items = await Promise.all(list.map((item) => cache.tryGet(item.link, async () => await getFullText(item))));
 
     return {
         title: $('title').text(),
